tests/wpt/tests/css/css-animations/parsing/animation-delay-end-invalid.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-delay-end">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("animation-delay-end", "infinite");
test_invalid_value("animation-delay-end", "0");
test_invalid_value("animation-delay-end", "1s 2s");
test_invalid_value("animation-delay-end", "1s / 2s");
test_invalid_value("animation-delay-end", "100px");
test_invalid_value("animation-delay-end", "100%");

test_invalid_value("animation-delay-end", "peek 50%");
test_invalid_value("animation-delay-end", "50% contain");
test_invalid_value("animation-delay-end", "50% cover");
test_invalid_value("animation-delay-end", "50% entry");
test_invalid_value("animation-delay-end", "50% enter");
test_invalid_value("animation-delay-end", "50% exit");
test_invalid_value("animation-delay-end", "contain contain");
test_invalid_value("animation-delay-end", "auto");
test_invalid_value("animation-delay-end", "none");
test_invalid_value("animation-delay-end", "cover 50% enter 50%");
test_invalid_value("animation-delay-end", "cover 100px");
test_invalid_value("animation-delay-end", "cover");
test_invalid_value("animation-delay-end", "contain");
test_invalid_value("animation-delay-end", "enter");
test_invalid_value("animation-delay-end", "exit");
</script>
